Clean product apis
