+++ src/Tweb.Modules/eTweb.Application/Catalog/Products/IProductService.cs
@@ -9,7 +9,7 @@ using System.Threading.Tasks;
 
 namespace eTweb.Application.Catalog.Products
 {
-    public interface IManageProductService
+    public interface IProductService
     {
         Task<int> Create(ProductCreateRequest request);
 
@@ -45,5 +45,7 @@ namespace eTweb.Application.Catalog.Products
         Task<ProductImageViewModel> GetImageById(int imageId);
 
         Task<List<ProductImageViewModel>> GetListImages(int productId);
+
+        Task<PagedResult<ProductViewModel>> GetAllByCategoryId(string languageId, GetPublicProductPagingRequest request);
     }
 }
