+++ src/Tweb.Modules/eTweb.Application/Catalog/Products/ProductService.cs
@@ -17,12 +17,12 @@ using eTweb.ViewModels.Catalog.ProductImages;
 
 namespace eTweb.Application.Catalog.Products
 {
-    public class ManageProductService : IManageProductService
+    public class ProductService : IProductService
     {
         private readonly eTwebDbContext _context;
         private readonly IStorageService _storageService;
 
-        public ManageProductService(
+        public ProductService(
             eTwebDbContext context,
             IStorageService storageService)
         {
@@ -333,6 +333,55 @@ namespace eTweb.Application.Catalog.Products
                     SortOrder = x.SortOrder
                 })
                 .ToListAsync();
+        }
+
+        public async Task<PagedResult<ProductViewModel>> GetAllByCategoryId(string languageId, GetPublicProductPagingRequest request)
+        {
+            // 1. Select join
+            var query = from p in _context.Products
+                        join pt in _context.ProductTranslations on p.Id equals pt.ProductId
+                        join pic in _context.ProductInCategories on p.Id equals pic.ProductId
+                        join c in _context.Categories on pic.CategoryId equals c.Id
+                        where pt.LanguageId == languageId
+                        select new { p, pt, pic };
+
+            // 2. Filter
+            if (request.CategoryId.HasValue && request.CategoryId.Value > 0)
+            {
+                query = query.Where(x => x.pic.CategoryId == request.CategoryId);
+            }
+
+            // 3. Paging
+            var totalRecord = await query.CountAsync();
+            var data = await query
+                .Skip((request.PageIndex - 1) * request.PageSize)
+                .Take(request.PageSize)
+                .Select(x => new ProductViewModel()
+                {
+                    Id = x.p.Id,
+                    Price = x.p.Price,
+                    OriginalPrice = x.p.OriginalPrice,
+                    Stock = x.p.Stock,
+                    ViewCount = x.p.ViewCount,
+                    DateCreated = x.p.DateCreated,
+                    Description = x.pt.Description,
+                    SeoDescription = x.pt.SeoDescription,
+                    Details = x.pt.Details,
+                    Name = x.pt.Name,
+                    LanguageId = x.pt.LanguageId,
+                    SeoAlias = x.pt.SeoAlias,
+                    SeoTitle = x.pt.SeoTitle
+                })
+                .ToListAsync();
+
+            // 4. Select and projection
+            var pagedResult = new PagedResult<ProductViewModel>()
+            {
+                Items = data,
+                TotalRecord = totalRecord
+            };
+
+            return pagedResult;
         }
     }
 }
