--- src/Tweb.Modules/eTweb.Application/Catalog/Products/PublicProductService.cs
@@ -1,101 +1,0 @@
-﻿using eTweb.Application.Dtos;
-using eTweb.Data.EF;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
-using eTweb.ViewModels.Catalog.Products;
-
-namespace eTweb.Application.Catalog.Products
-{
-    public class PublicProductService : IPubicProductService
-    {
-        private readonly eTwebDbContext _context;
-        public PublicProductService(eTwebDbContext context)
-        {
-            _context = context;
-        }
-
-        //public async Task<List<ProductViewModel>> GetAll(string languageId)
-        //{
-        //    var query = from p in _context.Products
-        //                join pt in _context.ProductTranslations on p.Id equals pt.ProductId
-        //                //join pic in _context.ProductInCategories on p.Id equals pic.ProductId
-        //                //join c in _context.Categories on pic.CategoryId equals c.Id
-        //                where pt.LanguageId == languageId
-        //                select new { p, pt/*, pic*/ };
-
-        //    var data = await query
-        //        .Select(x => new ProductViewModel()
-        //        {
-        //            Id = x.p.Id,
-        //            Price = x.p.Price,
-        //            OriginalPrice = x.p.OriginalPrice,
-        //            Stock = x.p.Stock,
-        //            ViewCount = x.p.ViewCount,
-        //            DateCreated = x.p.DateCreated,
-        //            Description = x.pt.Description,
-        //            SeoDescription = x.pt.SeoDescription,
-        //            Details = x.pt.Details,
-        //            Name = x.pt.Name,
-        //            LanguageId = x.pt.LanguageId,
-        //            SeoAlias = x.pt.SeoAlias,
-        //            SeoTitle = x.pt.SeoTitle
-        //        })
-        //        .ToListAsync();
-
-        //    return data;
-        //}
-
-        public async Task<PagedResult<ProductViewModel>> GetAllByCategoryId(string languageId, GetPublicProductPagingRequest request)
-        {
-            // 1. Select join
-            var query = from p in _context.Products
-                        join pt in _context.ProductTranslations on p.Id equals pt.ProductId
-                        join pic in _context.ProductInCategories on p.Id equals pic.ProductId
-                        join c in _context.Categories on pic.CategoryId equals c.Id
-                        where pt.LanguageId == languageId
-                        select new { p, pt, pic };
-
-            // 2. Filter
-            if (request.CategoryId.HasValue && request.CategoryId.Value > 0)
-            {
-                query = query.Where(x => x.pic.CategoryId == request.CategoryId);
-            }
-
-            // 3. Paging
-            var totalRecord = await query.CountAsync();
-            var data = await query
-                .Skip((request.PageIndex - 1) * request.PageSize)
-                .Take(request.PageSize)
-                .Select(x => new ProductViewModel()
-                {
-                    Id = x.p.Id,
-                    Price = x.p.Price,
-                    OriginalPrice = x.p.OriginalPrice,
-                    Stock = x.p.Stock,
-                    ViewCount = x.p.ViewCount,
-                    DateCreated = x.p.DateCreated,
-                    Description = x.pt.Description,
-                    SeoDescription = x.pt.SeoDescription,
-                    Details = x.pt.Details,
-                    Name = x.pt.Name,
-                    LanguageId = x.pt.LanguageId,
-                    SeoAlias = x.pt.SeoAlias,
-                    SeoTitle = x.pt.SeoTitle
-                })
-                .ToListAsync();
-
-            // 4. Select and projection
-            var pagedResult = new PagedResult<ProductViewModel>()
-            {
-                Items = data,
-                TotalRecord = totalRecord
-            };
-
-            return pagedResult;
-        }
-    }
-}
